Moved KitKat-specific code for external storage to separate file
